refactor: do basic auth in middleware, rather than a separate view
this removes the need for an extra URL and the redirect behaviour
